Move instance_filter function to own module
* Moved the instance_filter function from module_utils/morpheus_funcs to module_utils/instance_module_common module, resolves a circular import scenario
* Updated sanity test ignores
