+++ plugins/modules/instance_snapshot_info.py
@@ -95,11 +95,13 @@ from ansible.module_utils.connection import Connection
 try:
     import module_utils._info_module_common as info_module
     import module_utils.morpheus_funcs as mf
+    from module_utils.instance_module_common import instance_filter
     from module_utils.morpheusapi import MorpheusApi
     from module_utils._morpheus_classes import InstanceSnapshots
 except ModuleNotFoundError:
     import ansible_collections.morpheus.core.plugins.module_utils._info_module_common as info_module
     import ansible_collections.morpheus.core.plugins.module_utils.morpheus_funcs as mf
+    from ansible_collections.morpheus.core.plugins.module_utils.instance_module_common import instance_filter
     from ansible_collections.morpheus.core.plugins.module_utils.morpheusapi import MorpheusApi
     from ansible_collections.morpheus.core.plugins.module_utils._morpheus_classes import InstanceSnapshots
 
@@ -130,7 +132,7 @@ def run_module():
     connection = Connection(module._socket_path)
     morpheus_api = MorpheusApi(connection)
 
-    instances = mf.instance_filter(morpheus_api, module.params)
+    instances = instance_filter(morpheus_api, module.params)
 
     result['instance_snapshots'] = [
         mf.class_to_dict(InstanceSnapshots(
