+++ tests/sanity/ignore-2.11.txt
@@ -1,59 +1,3 @@
-plugins/module_utils/facts/appliance_database.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_database.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_database.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_database.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_database.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_database.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_database.py import-3.12 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_elastic.py import-3.12 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_license.py import-3.12 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_rabbitmq.py import-3.12 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_settings.py import-3.12 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_system.py import-3.12 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.6 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.7 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.8 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.9 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.10 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.11 # ignore circular import error
-plugins/module_utils/facts/appliance_threads.py import-3.12 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.6 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.7 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.8 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.9 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.10 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.11 # ignore circular import error
-plugins/module_utils/morpheusapi.py import-3.12 # ignore circular import error
 plugins/httpapi/morpheus.py validate-modules:missing-gplv3-license
 plugins/inventory/morpheus_inventory.py validate-modules:missing-gplv3-license
 plugins/lookup/search.py validate-modules:missing-gplv3-license
